test: Removed `test_acceptance.py` and integrated its contents in `test_main.py`
